feat: refactor AssetType handling and add PreviewAsset and VideoAsset classes
