+++ backend/src/FileService/FileService.Domain/AssetType.cs
@@ -6,14 +6,3 @@ public enum AssetType
     PREVIEW,
     AVATAR,
 }
-
-public static AssetType ToAssetType(this string value)
-{
-    return value.ToLowerInvariant() switch
-    {
-        "video" => AssetType.VIDEO,
-        "preview" => AssetType.PREVIEW,
-        "avatar" => AssetType.AVATAR,
-        _ => throw new ArgumentException($"Unknown AssetType: {value}"),
-    };
-}
