+++ backend/src/FileService/FileService.Domain/AssetTypeExtensions.cs
@@ -1,0 +1,15 @@
+namespace FileService.Domain;
+
+public static class AssetTypeExtensions
+{
+    public static AssetType ToAssetType(this string value)
+    {
+        return value.ToLowerInvariant() switch
+        {
+            "video" => AssetType.VIDEO,
+            "preview" => AssetType.PREVIEW,
+            "avatar" => AssetType.AVATAR,
+            _ => throw new ArgumentException($"Unknown AssetType: {value}"),
+        };
+    }
+}
